refactor : 로또 번호 체크 로직 내부에서 처리하도록 수정
