+++ src/test/java/lotto/model/LottoNumbersTest.java
@@ -18,9 +18,52 @@ class LottoNumbersTest {
     assertThat(new LottoNumbers(numberList).values()).isEqualTo(numberList);
   }
 
+  @ParameterizedTest
+  @MethodSource("numberListProvider")
+  void checkLottoNumberSize(List<Integer> numberList) {
+    new LottoNumbers(numberList);
+  }
+
+  @ParameterizedTest
+  @MethodSource("wrongSizeListProvider")
+  void checkLottoNumberSize_크기가6이아닌경우(List<Integer> numberList) {
+    assertThatExceptionOfType(IllegalArgumentException.class).isThrownBy(() -> {
+      new LottoNumbers(numberList);
+    });
+  }
+
+  @ParameterizedTest
+  @MethodSource("numberListProvider")
+  void checkLottoNumberRange(List<Integer> numberList) {
+    new LottoNumbers(numberList);
+  }
+
+  @ParameterizedTest
+  @MethodSource("wrongNumberListProvider")
+  void checkLottoNumberRange_Not1to45(List<Integer> numberList) {
+    assertThatExceptionOfType(IllegalArgumentException.class).isThrownBy(() -> {
+      new LottoNumbers(numberList);
+    });
+  }
+
   static Stream<Arguments> numberListProvider() {
     return Stream.of(
-        arguments(Arrays.asList(1, 2, 3, 4, 5, 6))
+        arguments(Arrays.asList(1, 2, 3, 4, 5, 6)),
+        arguments(Arrays.asList(1, 2, 3, 4, 5, 42))
+    );
+  }
+
+  static Stream<Arguments> wrongSizeListProvider() {
+    return Stream.of(
+        arguments(Arrays.asList(1, 2, 3, 4, 5)),
+        arguments(Arrays.asList(1, 2, 3, 4, 5, 6, 7))
+    );
+  }
+
+  static Stream<Arguments> wrongNumberListProvider() {
+    return Stream.of(
+        arguments(Arrays.asList(1, 2, 3, 4, 5, 46)),
+        arguments(Arrays.asList(0, 2, 3, 4, 5, 6))
     );
   }
 }
